update dependency
running kubefwd with an Azure user at a private cloud results in the
kubeconfig being made invalid. this triggers a re-login for kubectl
every time

the bug is traced down to the "k8s.io/client-go" dependency. upstream
fortunately has a bugfix:
https://github.com/kubernetes/client-go/commit/ba28f5cc8eb9738a322628953273802da189e165

this change bumps the version of the dependency to get the bugfix