Drop use of deprecated ioutils package
ioutils was deprecated in go 1.16. I have not seen plans for its full
removal, but that could happen at any point since all pre-deprecated
versions of Go are now end of life.

This is a trivial update to switch from using the deprecated module to
the current recommended replacement since the minimum kubefwd go version
is set to 1.18.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>